feat(chess): wire START_GAME/MAKE_MOVE/GET_BOARD/SURRENDER/HINT per-app actions (#9)
Expose Chess app logic to the vibe_action dispatch so Kayley can play chess
against Steven directly via MCP calls — even when the Chess window isn't open.

- Extract pure engine functions to apps/webuiapps/src/pages/Chess/engine.ts
  (unchanged behavior; Chess component imports them back). Adds boardToFen,
  fromNotation, legalMovesAlgebraic, and a simple suggestMove heuristic.
- Add handleChessAction() in vibeContainerMock.ts that reads/writes state.json
  (the same file the UI owns) using the engine module. When the window is
  open, each action is followed by a UI notification (AGENT_MOVE / NEW_GAME /
  SYNC_STATE) so the board refreshes from disk.
- Declare the 5 new actions in chess_en/meta.yaml with params + return shapes.

Build verified via pnpm build.
